Parse level secondary header.
